// resources/js/admin/game-maps/events/enums/game-map-move-event.ts
export enum GameMapMoveEvent {
  STATE_CHANGED = 'ADMIN_GAME_MAP_MOVE_STATE_CHANGED',
  START_LOCATION = 'ADMIN_GAME_MAP_MOVE_START_LOCATION',
  START_NPC = 'ADMIN_GAME_MAP_MOVE_START_NPC',
  SELECT_TARGET = 'ADMIN_GAME_MAP_MOVE_SELECT_TARGET',
  CONFIRM = 'ADMIN_GAME_MAP_MOVE_CONFIRM',
  CANCEL = 'ADMIN_GAME_MAP_MOVE_CANCEL',
  CLEAR_ERROR = 'ADMIN_GAME_MAP_MOVE_CLEAR_ERROR',
  SUCCEEDED = 'ADMIN_GAME_MAP_MOVE_SUCCEEDED',
}